Refactor validator move to middlewares
