+++ routes/userRouter.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const { userJoinValidator } = require("../validator/userValidator");
+const { userJoinValidator } = require("../middlewares/validator/userValidator");
 const { usersService } = require("../services/index");
 
 // user create에 대한 api
